refactor(e2e): split user activation tests into separate files
- Split monolithic 02-01-03-user-activation.test.ts into 2 files:
  - 02-01-03-01-user-activate.test.ts
  - 02-01-03-02-user-deactivate.test.ts
- Uses shared helpers from ui-helpers.ts (clickListAction, confirmYes,
  filterUsersList) instead of local duplicates
